move notifications to own partial
