+++ website/src/components/layout.js
@@ -1,6 +1,5 @@
 import React, { Fragment } from 'react';
 import Helmet from 'react-helmet';
-import classnames from 'classnames';
 import { graphql, StaticQuery } from 'gatsby';
 
 import Header from './header';
@@ -31,40 +30,18 @@ const Layout = ({ children }) => {
               }
             }
           }
-          notifs: file(relativePath: { regex: "/notifications/" }) {
-            childDataYaml {
-              notifications {
-                published
-                loud
-                message
-                url
-              }
-            }
-          }
         }
       `}
     >
       {data => {
         const { title, description } = data.site.siteMetadata;
-        const notifs = data.notifs.childDataYaml.notifications.filter(notif => notif.published);
 
         return (
           <Fragment>
             <Helmet defaultTitle={title} titleTemplate={`%s | ${title}`}>
               <meta name="description" content={description} />
             </Helmet>
-            {notifs.map((node, i) => (
-              <a
-                key={i}
-                href={node.url}
-                className={classnames('notification', {
-                  'notification-loud': node.loud,
-                })}
-              >
-                {node.message}
-              </a>
-            ))}
-            <Header notifications={notifs} />
+            <Header />
             {children}
             <Footer buttons={data.footer.childDataYaml.footer.buttons} />
           </Fragment>
